Fix for build issue with gaeDeploy.sh which fails, add softlink
Invocation of ./gradlew gaeUpload
fails with srcDirs path not found.  It appears the gradle-gae-plugin
(https://github.com/bmuschko/gradle-gae-plugin) has changed how it sets
both srcDirs and warDir with Gaelyk 1.2.  In commit: 3df134 on iUi,
the transition to Gaelyk 1.2 takes place.  This is a big deal and
affects the entire gaelyk build.  I believe, based on review of the
above github URL, that there is a bug in the gradle gae plugin that
always uses the default srcDirs path 'src/main/webapp' to find
WEB-INF/groovy.  I was unable to coerce the build.gradle to use
anything other than the expected default gradle directory structure.
So that it uses the default directory isn't a problem, but the problem
was that I couldn't override the srcDirs property.  I will explore this
later.

The quick fix is to add a soft link to 'src/main/webapp' ->../../web-app

This begs the question, why not adopt the proper gradle directory
structure?  So I see this fix as a workaround until we answer the real
question.